backend/requirements.txt
flask
gunicorn
pymongo[srv]
python-dotenv
passlib[bcrypt]
flask-cors